fix(web): compare cron bearer secrets in constant time
Every cron route is publicly reachable and CRON_SECRET is its only gate,
but five routes compared the Authorization header with plain string
inequality, leaking the secret through timing:

- /api/cron/vm-reconcile
- /api/cron/vm-alerts
- /api/cron/vm-reaper
- /api/cron/billing-reconcile
- /api/cron/indexnow

Two internal routes already did the timing-safe comparison inline
(/api/internal/vm/leases/revoke-expired, /api/internal/iroh/retention),
each with its own copy. This consolidates all seven onto one shared
authorizeCronRequest helper (services/cronAuth.ts) that trims, requires
the Bearer scheme, and compares with timingSafeEqual.

Each route keeps its exact fail-closed responses (the helper returns a
reason, not a Response), so no observable behavior changes for callers.
