+++ web/app/api/cron/indexnow/route.ts
@@ -1,13 +1,10 @@
+import { authorizeCronRequest } from "../../../../services/cronAuth";
 import { recentlyModifiedUrls, submitIndexNowUrls } from "../../../lib/indexnow";
 import sitemap from "../../../sitemap";
 
 
 export async function POST(request: Request): Promise<Response> {
-  const triggerSecret = process.env.CRON_SECRET?.trim();
-  if (
-    !triggerSecret ||
-    request.headers.get("authorization") !== `Bearer ${triggerSecret}`
-  ) {
+  if (!authorizeCronRequest(request).ok) {
     return Response.json({ error: "unauthorized" }, { status: 401 });
   }
