+++ web/app/api/cron/vm-alerts/route.ts
@@ -1,15 +1,14 @@
+import { authorizeCronRequest } from "../../../../services/cronAuth";
 import { runVmAlertChecks } from "../../../../services/observability/vmAlerts";
 import { jsonResponse } from "../../../../services/vms/routeHelpers";
 
 
 export async function GET(request: Request): Promise<Response> {
-  const cronSecret = process.env.CRON_SECRET?.trim();
-  if (!cronSecret) {
+  const auth = authorizeCronRequest(request);
+  if (!auth.ok && auth.reason === "cron_secret_missing") {
     return jsonResponse({ error: "cron_not_configured" }, 503);
   }
-
-  const expected = `Bearer ${cronSecret}`;
-  if (request.headers.get("authorization") !== expected) {
+  if (!auth.ok) {
     return jsonResponse({ error: "unauthorized" }, 401);
   }
