+++ web/app/api/internal/iroh/retention/route.ts
@@ -1,4 +1,3 @@
-import { timingSafeEqual } from "node:crypto";
 import * as Effect from "effect/Effect";
 import {
   IROH_RETENTION_MAX_DURATION_MS,
@@ -7,6 +6,7 @@ import {
   IrohRepositoryLive,
 } from "../../../../../services/iroh/repository";
 import { jsonResponse } from "../../../../../services/vms/routeHelpers";
+import { authorizeCronRequest } from "../../../../../services/cronAuth";
 
 
 export async function GET(request: Request): Promise<Response> {
@@ -18,15 +18,11 @@ export async function POST(request: Request): Promise<Response> {
 }
 
 async function handle(request: Request): Promise<Response> {
-  const secret = process.env.CRON_SECRET?.trim();
-  if (!secret) return jsonResponse({ error: "service_unavailable" }, 503);
-  const authorization = request.headers.get("authorization")?.trim() ?? "";
-  const token = authorization.toLowerCase().startsWith("bearer ")
-    ? authorization.slice("bearer ".length).trim()
-    : "";
-  const tokenBytes = Buffer.from(token);
-  const secretBytes = Buffer.from(secret);
-  if (tokenBytes.length !== secretBytes.length || !timingSafeEqual(tokenBytes, secretBytes)) {
+  const auth = authorizeCronRequest(request);
+  if (!auth.ok && auth.reason === "cron_secret_missing") {
+    return jsonResponse({ error: "service_unavailable" }, 503);
+  }
+  if (!auth.ok) {
     return jsonResponse({ error: "unauthorized" }, 401);
   }
