+++ web/app/api/internal/vm/leases/revoke-expired/route.ts
@@ -1,6 +1,6 @@
-import { timingSafeEqual } from "node:crypto";
 import { jsonResponse } from "@/services/vms/routeHelpers";
 import { revokeExpiredIdentityLeases, runVmWorkflow } from "@/services/vms/workflows";
+import { authorizeCronRequest } from "@/services/cronAuth";
 
 
 export async function GET(request: Request): Promise<Response> {
@@ -12,22 +12,12 @@ export async function POST(request: Request): Promise<Response> {
 }
 
 async function handle(request: Request): Promise<Response> {
-  const secret = process.env.CRON_SECRET?.trim();
-  if (!secret) {
+  const auth = authorizeCronRequest(request);
+  if (!auth.ok && auth.reason === "cron_secret_missing") {
     console.error("vm.leases.revoke_expired.cron_secret_missing");
     return jsonResponse({ error: "service_unavailable" }, 503);
   }
-
-  const authorization = request.headers.get("authorization")?.trim() ?? "";
-  const token = authorization.toLowerCase().startsWith("bearer ")
-    ? authorization.slice("bearer ".length).trim()
-    : "";
-  const tokenBuffer = Buffer.from(token);
-  const secretBuffer = Buffer.from(secret);
-  const tokenMatches =
-    tokenBuffer.length === secretBuffer.length &&
-    timingSafeEqual(tokenBuffer, secretBuffer);
-  if (!tokenMatches) {
+  if (!auth.ok) {
     return jsonResponse({ error: "unauthorized" }, 401);
   }
